+++ crates/rhai_impl/src/builtins.rs
@@ -44,7 +44,6 @@ pub fn register_all_widgets(
     register_primitive!("label", Label);
     register_primitive!("button", Button);
     register_primitive!("image", Image);
-    register_primitive!("icon", Icon);
     register_primitive!("input", Input);
     register_primitive!("progress", Progress);
     register_primitive!("combo_box_text", ComboBoxText);